user permissions

+++ posts/views.py
@@ -4,8 +4,11 @@ from .forms import PostForm
 from django.contrib import messages
 from django.core.paginator import Paginator, EmptyPage, PageNotAnInteger
 from urllib.parse import quote
+from django.http import Http404
 
 def post_create(request):
+	if not (request.user.is_staff or request.user.is_superuser):
+		raise Http404
 	form = PostForm(request.POST or None, request.FILES or None)
 	if form.is_valid():
 		form.save()
@@ -18,6 +21,8 @@ def post_create(request):
 	return render(request, 'post_create.html', context)
 
 def post_delete(request, post_slug):
+	if not (request.user.is_staff or request.user.is_superuser):
+		raise Http404
 	instance = get_object_or_404(Post, slug=post_slug)
 	instance.delete()
 	messages.warning(request, "Succesfully Deleted")
@@ -47,6 +52,8 @@ def post_list(request):
 	return render(request, 'post_list.html', context)
 
 def post_update(request, post_slug):
+	if not (request.user.is_staff or request.user.is_superuser):
+		raise Http404
 	instance = get_object_or_404(Post, slug=post_slug)
 	form = PostForm(request.POST or None, request.FILES or None, instance= instance)
 	if form.is_valid():
